FIx solution to use decimal to represent money, and store discounts as whole number percentage

+++ strategy_pattern/strategy_pattern/ICustomer.cs
@@ -5,6 +5,6 @@ namespace Strategy
     public interface ICustomer
     {
         bool IsMember();
-        double CalculatePrice(IProduct product);
+        Decimal CalculatePrice(IProduct product);
     }
 }
